Name the project Tuni, and make Etap 0 actually run
Renames every crate, the binary, the widget type, and the application ID.
Kero stays in the README as the macOS project whose behavior this one
reimplements, not as this project's name.

The spike now runs a real shell in a real window. Getting there fixed four
things the compiler and the first launch found:

- portable-pty has no arg0; new_default_prog is the constructor that starts a
  login shell, resolving $SHELL then the passwd entry and prefixing argv[0]
  with '-'. An explicit shell now travels through $SHELL, because the builder
  resolves the executable from argv[0] and a login-shell argv[0] is not a path.
  That deletes the hand-rolled lookup and the nix dependency with it.
- TerminalOptions is a root re-export of terminal::Options, and mouse
  positions are f32.
- The window title is a real GObject property now. Notifying a name that was
  never installed only produced a GLib critical.
- The application is NON_UNIQUE: a terminal launched from a shell has to
  inherit that shell's working directory, which a single primary instance
  could not see.

Pixel dimensions are saturating: cols times cell width overflows u16 on a wide
enough window, and a wrapped winsize would confuse every full-screen program.

Adds an offscreen capture path behind TUNI_CAPTURE_PNG, which renders the
widget through its own renderer and writes a PNG. This compositor has no
screenshot protocol, and a terminal that cannot be looked at cannot be
verified.

+++ crates/tuni-vt/src/lib.rs
@@ -19,7 +19,7 @@ use std::cell::RefCell;
 use std::rc::Rc;
 
 use libghostty_vt::render::{CellIterator, CursorVisualStyle, RenderState, RowIterator};
-use libghostty_vt::terminal::{ScrollViewport, Terminal as VtTerminal, TerminalOptions};
+use libghostty_vt::terminal::{Options as TerminalOptions, ScrollViewport, Terminal as VtTerminal};
 use libghostty_vt::{key, mouse};
 
 pub type Result<T> = std::result::Result<T, Error>;
@@ -180,8 +180,8 @@ impl Terminal {
             .set_button(input.button)
             .set_mods(input.mods)
             .set_position(mouse::Position {
-                x: f64::from(input.col),
-                y: f64::from(input.row),
+                x: f32::from(input.col),
+                y: f32::from(input.row),
             });
 
         self.encoded.clear();
